Cards context provider is a separate component
Refactor project structure
