+++ src/CardComponents/Card/Card.js
@@ -3,7 +3,7 @@ import cl from 'classnames';
 import './Card.scss';
 import CardHeader from './CardHeader/CardHeader';
 import CardBody from './CardBody/CardBody';
-import WithLoadingDelay from "../common/WithLoadingDelay/WithLoadingDelay";
+import WithLoadingDelay from "../../common/WithLoadingDelay/WithLoadingDelay";
 import PropTypes from 'prop-types';
 
 function Card({ className, onEdit,
